project/requirements.txt
azure_storage==0.37.0
InquirerPy==0.3.4
mysql_connector_repackaged==0.3.1
pandas==2.2.2
pandasgui==0.2.14
pyarrow==15.0.2
python-dotenv==1.0.1
tqdm==4.66.2
